+++ norbert/numpy/contrib.py
