+++ norbert/numpy/utils.py
